Compare message sources more correctly. `client.state.nick` is the nickname the client 'wants' and will request, but `client.upstream.state.nick` is their actual nickname, updated when the nickname is actually changed on the network.
However, we don't currently track the username in
`client.upstream.state.username` (e.g. when the ircd prepends a tilde to
the username, we don't pick that up and update that upstream state var),
so we won't bother comparing that property for now. Nick is the core
network identifier, so only comparing with that makes sense.